+++ src/config.ts
@@ -1,8 +1,6 @@
 import type { CsvSettings } from './formats/csv/CsvSettings';
 import type { IntermediateFilesStrategy } from './types/IntermediateFilesStrategy';
-import type { string_email } from './types/typeAliases';
-import type { string_name } from './types/typeAliases';
-import type { string_url_image } from './types/typeAliases';
+import type { string_email, string_name, string_url_image } from './types/typeAliases';
 import { just } from './utils/organization/just';
 
 /**
@@ -26,7 +24,7 @@ export const NAME = `Promptbook`;
  *
  * @public exported from `@promptbook/core`
  */
-export const ADMIN_EMAIL: string_email = 'me@pavolhejny.com';
+export const ADMIN_EMAIL: string_email = 'pavol@ptbk.io';
 
 /**
  * Name of the responsible person for the Promptbook on GitHub