+++ src/utils/validators/email/isValidEmail.test.ts
@@ -3,7 +3,7 @@ import { isValidEmail } from './isValidEmail';
 
 describe(`validation of emails`, () => {
     it(`is valid`, () => {
-        expect(isValidEmail(`me@pavolhejny.com`)).toBe(true);
+        expect(isValidEmail(`pavol@ptbk.io`)).toBe(true);
         expect(isValidEmail(`pavol@hejny.com`)).toBe(true);
         expect(isValidEmail(`dddd+aaaa@gmail.com`)).toBe(true);
         expect(isValidEmail(`strange-but-ok+email@webgpt.cz`)).toBe(true);